+++ src/server/game/Handlers/NPCHandler.cpp
@@ -298,7 +298,7 @@ void WorldSession::HandleTrainerBuySpellOpcode(WorldPacket& recvData)
         _player->LearnSpell(spellId, false);
 
     WorldPacket data(SMSG_TRAINER_BUY_SUCCEEDED, 12);
-    data << uint64(guid);
+    data << guid;
     data << uint32(spellId);
     SendPacket(&data);
 }
@@ -487,7 +487,7 @@ void WorldSession::SendBindPoint(Creature* npc)
     npc->CastSpell(_player, bindspell, true);
 
     WorldPacket data(SMSG_TRAINER_BUY_SUCCEEDED, 12);
-    data << uint64(npc->GetGUID());
+    data << npc->GetGUID();
     data << uint32(bindspell);
     SendPacket(&data);
 
@@ -536,7 +536,7 @@ void WorldSession::SendStablePetCallback(PreparedQueryResult result, ObjectGuid 
 
     WorldPacket data(MSG_LIST_STABLED_PETS, 200);           // guess size
 
-    data << uint64(guid);
+    data << guid;
 
     Pet* pet = _player->GetPet();
